Added page about the Postage Hike scheduled for December 3, 2017, renamed folder containing information on the March 12, 2017 Postage Hike, and created a short page in the old postage-hike-2017 directory with links to both pages.
Notes on possible problems: I'm not sure the path to the Postage Hike
image is correct, and am also not sure the links on the newly-created
short page will work as typed. Finally, I'm not at all certain whether
the newly-created page has an equal number of "div" and "/div" tags.